made breacking change, adding more algorithms to choose from
