+++ modules/net.plantabyte.drptrace.utils/src/net/plantabyte/drptrace/utils/ImageTracer.java
@@ -24,7 +24,7 @@ SOFTWARE.
 package net.plantabyte.drptrace.utils;
 
 import imagemagick.Quantize;
-import net.plantabyte.drptrace.Tracer;
+import net.plantabyte.drptrace.*;
 import net.plantabyte.drptrace.geometry.BezierShape;
 import net.plantabyte.drptrace.intmaps.ZOrderIntMap;
 
@@ -66,7 +66,7 @@ public class ImageTracer {
 				}
 			}
 		}
-		Tracer t = new Tracer();
+		IntervalTracer t = new IntervalTracer();
 		return t.traceAllShapes(ZOrderIntMap.fromMatrix(imgMatrix), smoothness);
 	}
 	/**
@@ -85,7 +85,7 @@ public class ImageTracer {
 	public static List<BezierShape> traceBufferedImage(
 			BufferedImage img, int smoothness
 	) {
-		Tracer t = new Tracer();
+		IntervalTracer t = new IntervalTracer();
 		return t.traceAllShapes(new BufferedImageIntMap(img), smoothness);
 	}
 }
